aula10/forum/requirements.txt
Django==2.2.4
djangorestframework==3.10.2
Markdown==3.1.1
pytz==2019.2
sqlparse==0.3.0
